product POST modified to insert taste, brewing method, and origin too

+++ backend/src/main/java/com/matchmycoffee/model/entity/ProductBrewingMethodId.java
@@ -18,5 +18,5 @@ public class ProductBrewingMethodId implements Serializable {
     private Long productId;
 
     @Column(name = "brewing_method_id")
-    private Integer brewingMethodId;
+    private Long brewingMethodId;
 }
